requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
Django==3.0.4
django-tenant-schemas==1.10.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
ordered-set==3.1.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wrapt==1.11.2
